refactor(runtime): bootstrap isolated worker entrypoints
