+++ .synergy/skill/integrate-llm/SKILL.md
@@ -63,7 +63,7 @@ Automatic reasoning variants are derived from model identity (`model.id`, API mo
 
 Production product inference enters `AgentTurn`: the Control Plane resolves final prompt and parameter plugin hooks plus serializable provider options into a request plan, request snapshots are schema-validated and capped, and the plan is sent as acknowledged chunks; event frames are bounded and acknowledged after consumption. The worker protocol owns its event projection: do not expose raw AI SDK stream objects as IPC types, and strip provider request bodies, response diagnostics, warnings, or other fields the Control Plane does not consume before checking the frame bound. Agent workers reconstruct built-in provider runtime functions without provider-plugin discovery. Keep executable callbacks, plugin runtimes and Host Services, session writers, permission promises, and other Control Plane handles out of the worker input. Model-facing tools are `ToolCatalog.Definition[]` only.
 
-The Agent worker runner's static value-import graph is also a product boundary. Keep Browser, Tool, Plugin, Plugin Runtime, `packages/browser`, and provider SDK implementations out of that graph. Resolve Control Plane preparation dependencies before transfer, and load only the selected built-in provider SDK lazily inside the worker. Update and run `test/session/agent-worker-runtime-boundary.test.ts` whenever the runner, LLM, provider, schema, or shared utility import graph changes.
+The executable bootstrap and Agent worker runner static value-import graphs are also product boundaries. Keep `src/index.ts` free of static application imports so compiled worker subcommands do not evaluate the main CLI/server graph. Keep Browser, Tool, Plugin, Plugin Runtime, `packages/browser`, and provider SDK implementations out of the Agent runner graph. Resolve Control Plane preparation dependencies before transfer, and load only the selected built-in provider SDK lazily inside the worker. Update and run `test/session/agent-worker-runtime-boundary.test.ts` whenever the executable entrypoint, runner, LLM, provider, schema, or shared utility import graph changes.
 
 Provider SSE protection is a per-event parser bound, not a total response, transport chunk, or process-memory limit. Keep code identifiers, error names, tests, and architecture documentation explicit about `SSE event parser bound` semantics. Test LF and CRLF event delimiters across chunk boundaries, including consecutive events exactly at the bound.
