+++ docs/architecture/execution-boundaries.md
@@ -34,7 +34,7 @@ The process boundary follows three ownership layers:
 - the elastic Agent worker pool owns provider inference and emits only projected model events and proposed tool calls;
 - tool runtimes own physical execution through their existing process, child-process, MCP, Browser, Link, or Control Plane transports.
 
-This split is a dependency boundary as well as an IPC boundary. The Agent worker runner's static import graph excludes Browser, Tool, Plugin, and Plugin Runtime implementations. Tool names, descriptions, and JSON Schemas cross into a worker; callbacks, Playwright/Chromium state, plugin processes, MCP clients, approval promises, and session writers do not. A model turn reaches its terminal provider result, disposes and releases the Agent worker, and only then can the Control Plane authorize and dispatch its proposed tools.
+This split is a dependency boundary as well as an IPC boundary. The compiled executable first enters a dependency-free dynamic bootstrap, so worker subcommands do not evaluate the main CLI/server graph. The Agent worker runner's static import graph excludes Browser, Tool, Plugin, and Plugin Runtime implementations. Tool names, descriptions, and JSON Schemas cross into a worker; callbacks, Playwright/Chromium state, plugin processes, MCP clients, approval promises, and session writers do not. A model turn reaches its terminal provider result, disposes and releases the Agent worker, and only then can the Control Plane authorize and dispatch its proposed tools.
 
 Policy workers isolate capability analysis from the HTTP/WebSocket event loop. Their protocol carries only the tool name, JSON-like arguments, and immutable workspace/plugin classification context. It bounds request size, queue depth, aggregate queued bytes, per-request time, IPC frames, request count, RSS, and heap use. Global-runtime startup begins prewarming without making HTTP/WebSocket availability depend on the child process; the first classification waits up to the fixed ten-second handshake deadline before the shorter per-request queue/transfer/classification deadline begins. Repeated pre-ready exits use exponential backoff and open a finite startup circuit instead of entering a respawn loop. The Control Plane remains the sole owner of profile compilation results, approval state, audit state, sandbox accumulation, and the final allow/ask/deny decision.
 
@@ -138,6 +138,7 @@ These restrictions are evaluated before the tool implementation. A permissive co
 - Every executable tool path passes through the centralized enforcement gate.
 - Model-facing tool definitions never contain executable callbacks.
 - Agent worker static imports never reach Browser, Tool, Plugin, or Plugin Runtime implementations.
+- The executable bootstrap has no static application imports and dynamically selects exactly one runtime entrypoint.
 - Permission decisions remain in the Control Plane and occur only after the Agent worker has released its turn.
 - Capability analysis runs in bounded Policy workers; those workers never decide authorization or execute tools.
 - Policy worker failure produces a finite conservative denial, never opens an approval wait, and cannot block HTTP/WebSocket service.
