+++ docs/architecture/llm-loop.md
@@ -271,7 +271,7 @@ Tool definitions cross the Agent worker boundary as plain JSON Schema data. AI S
 
 The external `AgentTurn` boundary transfers immutable turn snapshots through a versioned, schema-validated protocol. Before transfer, the Control Plane runs prompt and parameter plugin hooks, resolves the built-in provider profile, strips executable provider options, and freezes provider request timeouts into the worker plan. Requests are capped at 64 MiB and paged through acknowledged 1 MiB chunks rather than one unbounded IPC object. Worker event frames are Synergy-owned projections rather than raw AI SDK stream objects: lifecycle events retain only fields consumed by the Control Plane, so provider request bodies, response diagnostics, and warnings do not cross the process boundary. Frames are capped at 2 MiB, text/reasoning deltas coalesce up to 16 ms or 32 KiB, and each frame remains the only buffered frame until the Control Plane consumer acknowledges it.
 
-The worker runner has an enforced static runtime boundary. Its value-import graph may include the LLM loop, serializable schemas, provider reconstruction, and the selected provider SDK, but it may not include Browser, Tool, Plugin, Plugin Runtime, or `packages/browser` implementations. Provider SDKs load lazily after the worker receives a model plan; unselected SDKs do not enter a source-mode worker process. Browser names and JSON Schemas can cross as `ToolCatalog.Definition[]`, while Browser sessions, Playwright, tool callbacks, plugin discovery, Host Services, configuration writers, and permission state remain in the Control Plane or their owning tool runtime.
+The executable entrypoint is a thin bootstrap with no static application dependencies. It selects the main CLI/server, Agent runner, Policy runner, or Plugin runner through one mutually exclusive dynamic import. The Agent runner then has its own enforced static runtime boundary: its value-import graph may include the LLM loop, serializable schemas, provider reconstruction, and the selected provider SDK, but it may not include Browser, Tool, Plugin, Plugin Runtime, or `packages/browser` implementations. Provider SDKs load lazily after the worker receives a model plan; unselected SDKs do not enter a source-mode worker process. Browser names and JSON Schemas can cross as `ToolCatalog.Definition[]`, while Browser sessions, Playwright, tool callbacks, plugin discovery, Host Services, configuration writers, and permission state remain in the Control Plane or their owning tool runtime.
 
 The pool is elastic. `agentWorkers` is its parallelism ceiling; demand starts workers up to that ceiling instead of preallocating the full pool. `agentWorkerMinIdle` optionally maintains a warm reserve, and workers above the reserve retire after `agentWorkerIdleTimeoutMs`. Queue count and aggregate bytes are bounded independently. Cancellation, heartbeat timeout, startup circuit breaking, crash containment, RSS/heap/turn-count recycling, and parent-death cleanup affect only the owned turn.
