+++ packages/synergy/test/session/fixtures/agent-worker-runtime-boundary.ts
@@ -4,9 +4,7 @@ import ts from "typescript"
 const packageRoot = process.cwd()
 const workspaceRoot = path.resolve(packageRoot, "../..")
 const entrypoint = path.join(packageRoot, "src/session/agent-turn/runner.ts")
-const visited = new Set<string>()
-const parent = new Map<string, string>()
-const forbiddenExternalImports = new Set<string>()
+const bootstrapEntrypoint = path.join(packageRoot, "src/index.ts")
 
 function valueImports(source: ts.SourceFile): string[] {
   const imports: string[] = []
@@ -24,37 +22,47 @@ function valueImports(source: ts.SourceFile): string[] {
   return imports
 }
 
-async function visit(file: string): Promise<void> {
-  const normalized = path.normalize(file)
-  if (visited.has(normalized)) return
-  visited.add(normalized)
-  const source = ts.createSourceFile(
-    normalized,
-    await Bun.file(normalized).text(),
-    ts.ScriptTarget.Latest,
-    false,
-    normalized.endsWith(".tsx") ? ts.ScriptKind.TSX : ts.ScriptKind.TS,
-  )
-  for (const specifier of valueImports(source)) {
-    if (specifier.startsWith("@ai-sdk/") || specifier === "@openrouter/ai-sdk-provider") {
-      forbiddenExternalImports.add(specifier)
+async function collect(entry: string) {
+  const visited = new Set<string>()
+  const parent = new Map<string, string>()
+  const forbiddenExternalImports = new Set<string>()
+
+  async function visit(file: string): Promise<void> {
+    const normalized = path.normalize(file)
+    if (visited.has(normalized)) return
+    visited.add(normalized)
+    const source = ts.createSourceFile(
+      normalized,
+      await Bun.file(normalized).text(),
+      ts.ScriptTarget.Latest,
+      false,
+      normalized.endsWith(".tsx") ? ts.ScriptKind.TSX : ts.ScriptKind.TS,
+    )
+    for (const specifier of valueImports(source)) {
+      if (specifier.startsWith("@ai-sdk/") || specifier === "@openrouter/ai-sdk-provider") {
+        forbiddenExternalImports.add(specifier)
+      }
+      let resolved: string
+      try {
+        resolved = Bun.resolveSync(specifier, path.dirname(normalized))
+      } catch {
+        continue
+      }
+      if (!resolved.startsWith(workspaceRoot + path.sep) || resolved.includes(`${path.sep}node_modules${path.sep}`)) {
+        continue
+      }
+      if (!parent.has(resolved)) parent.set(resolved, normalized)
+      await visit(resolved)
     }
-    let resolved: string
-    try {
-      resolved = Bun.resolveSync(specifier, path.dirname(normalized))
-    } catch {
-      continue
-    }
-    if (!resolved.startsWith(workspaceRoot + path.sep) || resolved.includes(`${path.sep}node_modules${path.sep}`)) {
-      continue
-    }
-    if (!parent.has(resolved)) parent.set(resolved, normalized)
-    await visit(resolved)
   }
+
+  await visit(entry)
+  return { visited, parent, forbiddenExternalImports }
 }
 
-await visit(entrypoint)
-const inputs = [...visited].map((file) => path.relative(workspaceRoot, file))
+const runnerGraph = await collect(entrypoint)
+const bootstrapGraph = await collect(bootstrapEntrypoint)
+const inputs = [...runnerGraph.visited].map((file) => path.relative(workspaceRoot, file))
 const forbidden = inputs.filter(
   (input) =>
     input.startsWith("packages/synergy/src/browser/") ||
@@ -63,14 +71,14 @@ const forbidden = inputs.filter(
     input.startsWith("packages/synergy/src/tool/") ||
     input.startsWith("packages/browser/"),
 )
-forbidden.push(...[...forbiddenExternalImports].map((specifier) => `external:${specifier}`))
+forbidden.push(...[...runnerGraph.forbiddenExternalImports].map((specifier) => `external:${specifier}`))
 
 const dependencyPath = (relative: string) => {
   const result: string[] = []
   let current: string | undefined = path.join(workspaceRoot, relative)
   while (current) {
     result.unshift(path.relative(workspaceRoot, current))
-    current = parent.get(current)
+    current = runnerGraph.parent.get(current)
   }
   return result
 }
@@ -79,8 +87,11 @@ console.log(
   JSON.stringify({
     success: true,
     logs: [],
-    entryFound: visited.has(entrypoint),
+    entryFound: runnerGraph.visited.has(entrypoint),
     forbidden,
     firstForbiddenPath: forbidden[0] ? dependencyPath(forbidden[0]) : [],
+    bootstrapStaticDependencies: [...bootstrapGraph.visited]
+      .filter((file) => file !== bootstrapEntrypoint)
+      .map((file) => path.relative(workspaceRoot, file)),
   }),
 )
